tools/etl/requirements.txt
requests>=2.31.0
geopandas>=0.14.0
pandas>=2.0.0
python-dotenv>=1.0.0
fiona>=1.9.5
shapely>=2.0.0

